Fixed bugs and added flexibility in buildfeastini
Corrected a typo where the ROMS grid name should be passed as an
input rather than hardcoded.

Also, added an additional optional input flag specifying whether
to call the add-data R script from within the subfunction.  The
default is true (same as old behavior), but this allows one to
add the data in a more manual way if necessary (in today's case,
I needed to add 2004 data to a 2038 file; the current script expects
the input and output file names to match, so had to do a little
trickery before running the R script).