Split runroms into tow functions (createinputfiles and runroms)
I've now split the old runroms function into two parts.  The
new creatinputfiles function handles all the details of creating
the input and output files and folders necessary for a ROMS
simulation.  The new runroms handles only the task of making the
call to mpirun with the appropriate ROMS executable, input,
standard output, and standard error files.

This change allow the code to be more flexible across machines,
which may need cusomized methods to actually call mpirun (i.e
on machines with schedulers).