Updates to moxroms to account for new input file organization scheme
Input files underwent a bit of an upgrade.  We now keep both CORE and
CFS files around for an overlapping period (1005-2003 right now).  So
the setinputfiles function now takes an optional parameter (breakyr)
that indicates when to switch from one dataset to another.

Also added a bit to check how many layers are being used and to choose
boundary files from the appropriate folder.

Also made some changes to the basic naming scheme for all surface
forcing and boundary files... that's just for consistency with Al's
naming scheme, which is more clear than my old one.