Added setvinfo option to moxroms.runhindcast and setinfiles
In the original version of setinfiles, the varinfo.dat file name
is set to the most recent version used for BESTNPZ runs.  In order
to allow flexibility to run other version (e.g. ACLIM runs) that
require a different varinfo, I've now added a flag that indicates
whether the VARNAME parameter should be overwritten or not.  Still
back-compatible with older run scripts (not specifying defaults to
True, i.e. overwrite the value).